# openapi.yaml
openapi: 3.0.3
info:
  title: Plumbing JMS API
  version: 1.0.0
  description: API for the Plumbing Job Management System
servers:
  - url: http://localhost:3000/api
    description: Development server
paths:
...
  /invoices:
    get:
      summary: Get invoices
      parameters:
        - in: query
          name: customerId
          schema:
            type: string
          description: Filter by customer ID
      responses:
        '200':
          description: A list of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invoice'
    post:
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Invoice'
      responses:
        '201':
          description: Invoice created
...
  /procurement:
    get:
      summary: Get procurement items
      responses:
        '200':
          description: A list of procurement items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProcurementItem'
    post:
      summary: Create a procurement item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcurementItem'
      responses:
        '201':
          description: Procurement item created
...
components:
  schemas:
    Job:
      type: object
      properties:
        id:
          type: string
        jobNumber:
          type: string
        customerId:
          type: string
        description:
          type: string
        adminHourlyRate:
          type: number
          format: float
        adminNotes:
          type: string
        assignedTo:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled, assigned, to_be_invoiced, invoiced]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type: string
        labourCost:
          type: number
          format: float
          default: 0
        materialsCost:
          type: number
          format: float
          default: 0
        subtotal:
          type: number
          format: float
          default: 0
        vatAmount:
          type: number
          format: float
          default: 0
        grandTotal:
          type: number
          format: float
          default: 0
        customer:
          type: object
          properties:
            name:
              type: string
            email:
              type: string
            phone:
              type: string
        assignedToProfile:
          type: object
          properties:
            id:
              type: string
            fullName:
              type: string
            email:
              type: string
...
    Invoice:
      type: object
      properties:
        id:
          type: string
        jobCardId:
          type: string
        customerId:
          type: string
        invoiceNumber:
          type: string
        amountDue:
          type: number
          format: float
        vatAmount:
          type: number
          format: float
        amountPaid:
          type: number
          format: float
          default: 0
        status:
          type: string
          enum: [unpaid, partial, paid]
        dueDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        job:
          $ref: '#/components/schemas/Job'
        customer:
          type: object
          properties:
            name:
              type: string
            email:
              type: string
            phone:
              type: string
        payments:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              amount:
                type: number
                format: float
              method:
                type: string
                enum: [cash, card, bank_transfer, check]
              note:
                type: string
              createdAt:
                type: string
                format: date-time
              recordedBy:
                type: string
...
    ProcurementItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        quantityOnHand:
          type: integer
          minimum: 0
        unit:
          type: string
        costPerUnit:
          type: number
          format: float
        supplier:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...